+++ server/zanata-frontend/src/app/editor/components/PhraseStatusFilter/PhraseStatusFilter.test.js
@@ -165,6 +165,7 @@ describe('PhraseStatusFilterTest', () => {
     }
 
     let filterChangeType = 'none'
+    // @ts-ignore any
     const onFilterChange = statusType => {
       filterChangeType = statusType
     }