+++ server/zanata-frontend/src/app/editor/components/Suggestion.js
@@ -40,6 +40,7 @@ class Suggestion extends React.Component {
    * Calculate the match type for the suggestion
    * @type Function
    */
+  // @ts-ignore any
   matchType = (suggestion) => {
     let topMatch = suggestion.matchDetails[0]